Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
-float main() {
+int main() {
     float num;
     scanf("%f",&num);
-    printf("You entered: %f",num);
+    printf("You entered: %.2f",num);
     return 0;
 }
